restyling + change of logo
Difference in background color  -  The 'graph-demo' div does not really look well with the mockup bcgrd color. Might go with the white.

Labels for the parameter and graph - I kept it to be more informative to users.

Change in intro - since the R app does not require the additional step to load the local version, it has been adjusted.