feat: add new deploy script
This new script handles deployment of both Python and R Dash apps within a monorepo. At the moment, it will skip any apps that require redis, and also detects when the old-style monorepo was deployed in order to override the repository with the newer, slimmer methodology.

This script is meant to be run from CI, and should be avoided if possible locally, though will still work, assuming the dds-client is available.


Former-commit-id: 21c54a9e3fe536fadd3a04cd3042bc9ea2378ed6