refactor: remove need to reference apps/APP subdirectory across codebase
This made it more complex for others to reference the code - they'd need to remove the paths themselves - and also complicated our deployment process.

With the new deploy process, we treat each app directory as if it was a single entity, using a temporary git repository if necessary to initialize the app, and update the app with history as appropriate.


Former-commit-id: 64234735c29d42e4e01c684718d51f414970b597