fix funny shapes or else dash.no_update
If users draw the rectangle upside-down, the y values get switched and
the array indices are no longer ascending, so now the y values are
always sorted before determining the top and bottom of the rectangle.

If users draw shapes that are too small in any of the views or
histogram, then dash.no_update is returned.